Cleanup av pakkestruktur og kompakter gravid søknad kvittering
