+++ src/main/kotlin/no/nav/helse/fritakagp/integration/altinn/kvittering/gravid/soeknad/GravidSoeknadKvitteringProcessor.kt
@@ -1,4 +1,4 @@
-package no.nav.helse.fritakagp.processing.kvittering
+package no.nav.helse.fritakagp.integration.altinn.kvittering.gravid.soeknad
 
 import com.fasterxml.jackson.databind.ObjectMapper
 import no.nav.helse.arbeidsgiver.bakgrunnsjobb.BakgrunnsjobbProsesserer
@@ -7,14 +7,14 @@ import java.lang.IllegalArgumentException
 import java.time.LocalDateTime
 import java.util.*
 
-class KvitteringProcessor(
-        val kvitteringSender: KvitteringSender,
-        val db: GravidSoeknadRepository,
-        val om: ObjectMapper
+class GravidSoeknadKvitteringProcessor(
+    private val gravidSoeknadKvitteringSender: GravidSoeknadKvitteringSender,
+    private val db: GravidSoeknadRepository,
+    private val om: ObjectMapper
 ) : BakgrunnsjobbProsesserer {
 
     companion object {
-        val JOB_TYPE = "kvittering"
+        val JOB_TYPE = "kvitt_søk_gravid"
     }
 
     override fun nesteForsoek(forsoek: Int, forrigeForsoek: LocalDateTime): LocalDateTime {
@@ -22,14 +22,14 @@ class KvitteringProcessor(
     }
 
     override fun prosesser(jobbData: String) {
-        val kvitteringJobbData = om.readValue(jobbData, KvitteringJobData::class.java)
+        val kvitteringJobbData = om.readValue(jobbData, Jobbdata::class.java)
         val soeknad = db.getById(kvitteringJobbData.soeknadId)
             ?: throw IllegalArgumentException("Fant ikke søknaden i jobbdatanene $jobbData")
 
-        kvitteringSender.send(soeknad)
+        gravidSoeknadKvitteringSender.send(soeknad)
     }
-}
 
-data class KvitteringJobData(
+    data class Jobbdata(
         val soeknadId: UUID
-)
+    )
+}
