+++ src/main/kotlin/no/nav/helse/fritakagp/integration/oauth2/OAuthClientConfig.kt
@@ -1,4 +1,4 @@
-package no.nav.helse.fritakagp.oauth2
+package no.nav.helse.fritakagp.integration.oauth2
 
 import com.nimbusds.oauth2.sdk.auth.ClientAuthenticationMethod
 import io.ktor.config.*
