+++ src/main/kotlin/no/nav/helse/fritakagp/koin/KoinProfiles.kt
@@ -22,12 +22,10 @@ import no.nav.helse.arbeidsgiver.integrasjoner.oppgave.OpprettOppgaveRequest
 import no.nav.helse.arbeidsgiver.integrasjoner.oppgave.OpprettOppgaveResponse
 import no.nav.helse.arbeidsgiver.integrasjoner.pdl.*
 import no.nav.helse.arbeidsgiver.kubernetes.KubernetesProbeManager
-import no.nav.helse.fritakagp.processing.kvittering.DummyKvitteringSender
-import no.nav.helse.fritakagp.processing.kvittering.KvitteringSender
 import no.nav.helse.fritakagp.gcp.BucketStorage
 import no.nav.helse.fritakagp.gcp.MockBucketStorage
-import no.nav.helse.fritakagp.virusscan.MockVirusScanner
-import no.nav.helse.fritakagp.virusscan.VirusScanner
+import no.nav.helse.fritakagp.integration.virusscan.MockVirusScanner
+import no.nav.helse.fritakagp.integration.virusscan.VirusScanner
 import org.koin.core.Koin
 import org.koin.core.definition.Kind
 import org.koin.core.module.Module
