+++ src/main/kotlin/no/nav/helse/fritakagp/koin/LocalKoinProfile.kt
@@ -11,9 +11,9 @@ import no.nav.helse.fritakagp.processing.gravid.GravidSoeknadPDFGenerator
 import no.nav.helse.fritakagp.processing.gravid.SoeknadGravidProcessor
 import no.nav.helse.fritakagp.processing.kronisk.KroniskSoeknadPDFGenerator
 import no.nav.helse.fritakagp.processing.kronisk.SoeknadKroniskProcessor
-import no.nav.helse.fritakagp.processing.kvittering.DummyKvitteringSender
-import no.nav.helse.fritakagp.processing.kvittering.KvitteringProcessor
-import no.nav.helse.fritakagp.processing.kvittering.KvitteringSender
+import no.nav.helse.fritakagp.integration.altinn.kvittering.gravid.soeknad.DummyGravidSoeknadKvitteringSender
+import no.nav.helse.fritakagp.integration.altinn.kvittering.gravid.soeknad.GravidSoeknadKvitteringProcessor
+import no.nav.helse.fritakagp.integration.altinn.kvittering.gravid.soeknad.GravidSoeknadKvitteringSender
 import org.koin.dsl.bind
 import org.koin.dsl.module
 import javax.sql.DataSource
@@ -34,6 +34,6 @@ fun localDevConfig(config: ApplicationConfig) = module {
     single { SoeknadGravidProcessor(get(), get(), get(), get(), GravidSoeknadPDFGenerator(), get(), get())}
     single { SoeknadKroniskProcessor(get(), get(), get(), get(), KroniskSoeknadPDFGenerator(), get(), get())}
 
-    single { DummyKvitteringSender() } bind KvitteringSender::class
-    single { KvitteringProcessor(get(), get(), get())}
+    single { DummyGravidSoeknadKvitteringSender() } bind GravidSoeknadKvitteringSender::class
+    single { GravidSoeknadKvitteringProcessor(get(), get(), get()) }
 }
