--- src/main/kotlin/no/nav/helse/fritakagp/processing/kvittering/KvitteringSender.kt
@@ -1,7 +1,0 @@
-package no.nav.helse.fritakagp.processing.kvittering
-
-import no.nav.helse.fritakagp.domain.SoeknadGravid
-
-interface KvitteringSender {
-    fun send(kvittering: SoeknadGravid)
-}
